+++ consensus/types/errors.go
@@ -1,5 +1,7 @@
 package types
 
+// TECHDEBT: Avoid having a centralized file for all errors (harder to maintain and identify).
+
 import (
 	"encoding/base64"
 	"errors"
@@ -19,12 +21,12 @@ const (
 	ProposalBlockExtends     = "the ProposalQC block is the same as the LockedQC block"
 
 	// WARN
-	NilUtilityContextWarning     = "[WARN] Utility context not nil when preparing a new block? Releasing for now but should not happen"
-	InvalidPartialSigInQCWarning = "[WARN] QC contains an invalid partial signature"
+	NilUtilityContextWarning     = "⚠️ [WARN] utilityContext expected to be nil but is not. TODO: Investigate why this is and fix it"
+	InvalidPartialSigInQCWarning = "⚠️ [WARN] QC contains an invalid partial signature"
 
 	// DEBUG
-	DebugResetToGenesis  = "[DEBUG] Resetting to genesis..."
-	DebugTriggerNextView = "[DEBUG] Triggering next view..."
+	DebugResetToGenesis  = "🧑‍💻 [DEVELOP] Resetting to genesis..."
+	DebugTriggerNextView = "🧑‍💻 [DEVELOP] Triggering next view..."
 )
 
 var StepToString map[HotstuffStep]string
@@ -36,44 +38,49 @@ func init() {
 	}
 }
 
-func PacemakerInterrupt(height uint64, step HotstuffStep, round uint64) string {
-	return fmt.Sprintf("INTERRUPT at (height, step, round): (%d, %s, %d)!", height, StepToString[step], round)
+// TODO(#288): Improve all of this logging:
+// 1. Replace `fmt.Sprintf` with `log.Printf` (or similar)
+// 2. Add the appropriate log level (warn, debug, etc...) where appropriate
+// 3. Remove this file and move log related text into place (easier to maintain, debug, understand, etc.)
+
+func PacemakerInterrupt(reason string, height uint64, step HotstuffStep, round uint64) string {
+	return fmt.Sprintf("⏰ Interrupt ⏰ at (height, step, round): (%d, %s, %d)! Reason: %s", height, StepToString[step], round, reason)
 }
 
 func PacemakerTimeout(height uint64, step HotstuffStep, round uint64) string {
-	return fmt.Sprintf("Timed out at (height, step, round) (%d, %s, %d)!", height, StepToString[step], round)
+	return fmt.Sprintf("⌛ Timed out ⌛ at (height, step, round) (%d, %s, %d)!", height, StepToString[step], round)
 }
 
 func PacemakerNewHeight(height uint64) string {
-	return fmt.Sprintf("Starting first round for new block at height: %d", height)
+	return fmt.Sprintf("🏁 Starting 1st round 🏁 for height: %d", height)
 }
 
 func PacemakerCatchup(height1, step1, round1, height2, step2, round2 uint64) string {
-	return fmt.Sprintf("pacemaker catching up the node's (height, step, round) FROM (%d, %s, %d) TO (%d, %s, %d)", height1, StepToString[HotstuffStep(step1)], round1, height2, StepToString[HotstuffStep(step2)], round2)
+	return fmt.Sprintf("🏃 Pacemaker catching 🏃 up (height, step, round) FROM (%d, %d, %d) TO (%d, %d, %d)", height1, step1, round1, height2, step2, round2)
 }
 
 func OptimisticVoteCountWaiting(step HotstuffStep, status string) string {
-	return fmt.Sprintf("Still waiting for more %s messages; %s", StepToString[step], status)
+	return fmt.Sprintf("⏳ Waiting ⏳for more %s messages; %s", StepToString[step], status)
 }
 
-func OptimisticVoteCountPassed(step HotstuffStep) string {
-	return fmt.Sprintf("received enough %s votes!", StepToString[step])
+func OptimisticVoteCountPassed(height uint64, step HotstuffStep, round uint64) string {
+	return fmt.Sprintf("📬 Received enough 📬 votes at (height, step, round) (%d, %s, %d)", height, StepToString[step], round)
 }
 
 func CommittingBlock(height uint64, numTxs int) string {
-	return fmt.Sprintf("🧱🧱🧱 Committing block at height %d with %d transactions 🧱🧱🧱", height, numTxs)
+	return fmt.Sprintf("🧱 Committing block 🧱 at height %d with %d transactions", height, numTxs)
 }
 
 func ElectedNewLeader(address string, nodeId NodeId, height, round uint64) string {
-	return fmt.Sprintf("👑 Elected new leader for (%d-%d): %d (%s) 👑", height, round, nodeId, address)
+	return fmt.Sprintf("🙇 Elected leader 🙇 for height/round %d/%d: [%d] (%s)", height, round, nodeId, address)
 }
 
 func ElectedSelfAsNewLeader(address string, nodeId NodeId, height, round uint64) string {
-	return fmt.Sprintf("👑👑👑👑👑👑 I am the new leader for (%d-%d): %d (%s) 👑👑👑👑👑👑👑👑", height, round, nodeId, address)
+	return fmt.Sprintf("👑 I am the leader 👑 for height/round %d/%d: [%d] (%s)", height, round, nodeId, address)
 }
 
 func SendingMessage(msg *HotstuffMessage, nodeId NodeId) string {
-	return fmt.Sprintf("Sending %s message to %d", StepToString[msg.GetStep()], nodeId)
+	return fmt.Sprintf("✉️ Sending message ✉️ to %d at (height, step, round) (%d, %d, %d)", nodeId, msg.Height, msg.Step, msg.Round)
 }
 
 // TODO consider putting a better value for logs
@@ -82,7 +89,11 @@ func SendingStateSyncMessage(msg *anypb.Any, nodeId NodeId) string {
 }
 
 func BroadcastingMessage(msg *HotstuffMessage) string {
-	return fmt.Sprintf("Broadcasting message for %s step", StepToString[msg.GetStep()])
+	return fmt.Sprintf("📣 Broadcasting message 📣 (height, step, round): (%d, %d, %d)", msg.GetHeight(), msg.GetStep(), msg.GetRound())
+}
+
+func RestartTimer() string {
+	return fmt.Sprintln("Restarting timer")
 }
 
 func WarnInvalidPartialSigInQC(address string, nodeId NodeId) string {
@@ -90,32 +101,37 @@ func WarnInvalidPartialSigInQC(address string, nodeId NodeId) string {
 }
 
 func WarnMissingPartialSig(msg *HotstuffMessage) string {
-	return fmt.Sprintf("[WARN] No partial signature found for step %s which should not happen...", StepToString[msg.GetStep()])
+	return fmt.Sprintf("⚠️ [WARN] No partial signature found for step %s which should not happen...", StepToString[msg.GetStep()])
 }
 
 func WarnDiscardHotstuffMessage(_ *HotstuffMessage, reason string) string {
-	return fmt.Sprintf("[WARN] %s because: %s", DisregardHotstuffMessage, reason)
+	return fmt.Sprintf("⚠️ [WARN] %s because: %s", DisregardHotstuffMessage, reason)
 }
 
 func WarnUnexpectedMessageInPool(_ *HotstuffMessage, height uint64, step HotstuffStep, round uint64) string {
-	return fmt.Sprintf("[WARN] Message in pool does not match (height, step, round) of QC being generated; %d, %s, %d", height, StepToString[step], round)
+	return fmt.Sprintf("⚠️ [WARN] Message in pool does not match (height, step, round) of QC being generated; %d, %s, %d", height, StepToString[step], round)
 }
 
 func WarnIncompletePartialSig(ps *PartialSignature, msg *HotstuffMessage) string {
-	return fmt.Sprintf("[WARN] Partial signature is incomplete for step %s which should not happen...", StepToString[msg.GetStep()])
+	return fmt.Sprintf("⚠️ [WARN] Partial signature is incomplete for step %s which should not happen...", StepToString[msg.GetStep()])
 }
 
 func DebugTogglePacemakerManualMode(mode string) string {
-	return fmt.Sprintf("[DEBUG] Toggling pacemaker manual mode to %s", mode)
+	return fmt.Sprintf("🔎 [DEBUG] Toggling pacemaker manual mode to %s", mode)
 }
 
 func DebugNodeState(state ConsensusNodeState) string {
-	return fmt.Sprintf("[DEBUG] NODE STATE: Node %d is at (Height, Step, Round): (%d, %d, %d)\n", state.NodeId, state.Height, state.Step, state.Round)
+	return fmt.Sprintf("🔎 [DEBUG] Node %d is at (Height, Step, Round): (%d, %d, %d)", state.NodeId, state.Height, state.Step, state.Round)
 }
 
+// TODO(olshansky): Add source and destination NodeId of message here
+func DebugReceivedHandlingHotstuffMessage(msg *HotstuffMessage) string {
+	return fmt.Sprintf("🔎 [DEBUG] Received hotstuff msg at (Height, Step, Round): (%d, %d, %d)", msg.Height, msg.GetStep(), msg.Round)
+}
+
+// TODO(olshansky): Add source and destination NodeId of message here
 func DebugHandlingHotstuffMessage(msg *HotstuffMessage) string {
-	// TODO(olshansky): Add source and destination NodeId of message here
-	return fmt.Sprintf("[DEBUG] Handling message w/ Height: %d; Type: %s; Round: %d.", msg.Height, StepToString[msg.GetStep()], msg.Round)
+	return fmt.Sprintf("🔎 [DEBUG] Handling hotstuff msg at (Height, Step, Round): (%d, %d, %d)", msg.Height, msg.GetStep(), msg.Round)
 }
 
 // Errors
@@ -158,6 +174,8 @@ const (
 	createStateSyncMessageError                 = "error creating state sync message"
 	anteValidationError                         = "discarding hotstuff message because ante validation failed"
 	nilLeaderIdError                            = "attempting to send a message to leader when LeaderId is nil"
+	newPersistenceReadContextError              = "error creating new persistence read context"
+	persistenceGetAllValidatorsError            = "error getting all validators from persistence"
 )
 
 var (
@@ -194,6 +212,8 @@ var (
 	ErrCreateStateSyncMessage                 = errors.New(createStateSyncMessageError)
 	ErrHotstuffValidation                     = errors.New(anteValidationError)
 	ErrNilLeaderId                            = errors.New(nilLeaderIdError)
+	ErrNewPersistenceReadContext              = errors.New(newPersistenceReadContextError)
+	ErrPersistenceGetAllValidators            = errors.New(persistenceGetAllValidatorsError)
 )
 
 func ErrInvalidBlockSize(blockSize, maxSize uint64) error {